requirements.txt
# Multi-agent framework
crewai==1.12.2

# LLM provider
openai>=1.83.0

# arXiv API + PDF parsing
arxiv==2.4.1
PyMuPDF>=1.24.0

# Web UI
streamlit>=1.39.0

# Token counting
tiktoken>=0.7.0

# PDF report generation
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
